+++ src/internal/adapter/THEOplayerAdapter.ts
@@ -15,6 +15,7 @@ import type {
   PresentationModeChangeEvent,
   ProgressEvent,
   RateChangeEvent,
+  ResizeEvent,
   SourceDescription,
   TextTrack,
   TextTrackEvent,
@@ -78,6 +79,8 @@ const defaultPlayerState: NativePlayerState = {
   selectedVideoTrack: undefined,
   selectedAudioTrack: undefined,
   selectedTextTrack: undefined,
+  width: undefined,
+  height: undefined,
 };
 
 export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> implements THEOplayer {
@@ -114,6 +117,7 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
     this.addEventListener(PlayerEventType.MEDIA_TRACK, this.onMediaTrack);
     this.addEventListener(PlayerEventType.MEDIA_TRACK_LIST, this.onMediaTrackList);
     this.addEventListener(PlayerEventType.PRESENTATIONMODE_CHANGE, this.onPresentationModeChange);
+    this.addEventListener(PlayerEventType.RESIZE, this.onResize);
   }
 
   private onSourceChange = () => {
@@ -141,6 +145,11 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
 
   private onTimeupdate = (event: TimeUpdateEvent) => {
     this._state.currentTime = event.currentTime;
+  };
+
+  private onResize = (event: ResizeEvent) => {
+    this._state.width = event.width;
+    this._state.height = event.height;
   };
 
   private onLoadedMetadata = (event: LoadedMetadataEvent) => {
@@ -295,7 +304,7 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
   }
 
   get seekable() {
-    return this._state.seekable;
+    return this._state.seekable ?? [];
   }
 
   private seekableEnd(): number {
@@ -304,7 +313,7 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
   }
 
   get buffered() {
-    return this._state.buffered;
+    return this._state.buffered ?? [];
   }
 
   get cast(): CastAPI {
@@ -558,5 +567,13 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
       Object.assign(this._state, state);
     }
     this._castAdapter.init_();
+  }
+
+  get width(): number | undefined {
+    return this._state.width;
+  }
+
+  get height(): number | undefined {
+    return this._state.height;
   }
 }